Add hidden support for 'userid' parameter on login
This parameter allows the username field to be auto-filled (as a
suggestion) by the WAA when directing a user to the login page.  The
user can still amend the field as they wish, contrary to the existing
'force_userid' template variable used when obtaining active consent or
re-authentication from an already logged-in user.

This matches easter-egg behaviour observed on UIS Raven:

    https://raven.cam.ac.uk/auth/authenticate.html?ver=1&url=https://raven.cam.ac.uk/debug.html&userid=spqr2

Due to a new reliance on Flask's (actually Werkzeug's) request.args
property, rather than built-in query string processing based on the
Python standard library as part of ucam-wls, this commit also removes
support for query strings using a semicolon separator ';' rather than
the more conventional ampersand '&'.  I think this is fine in practice.

Signed-off-by: Edwin Balani <git@balani.xyz>